Fix auth
